Implemented a Type Checker for all statements and expressions
The type checker is called whenever a new example is created in the Interpreter class and if no exceptions are thrown, the example is added to the menu. This type checker's main purpose is to ensure that the examples created are made as the toy language expects them to be made before actually executing them at runtime.